+++ src/app/app.component.ts
@@ -1,16 +1,9 @@
-import {AfterViewInit, Component, ElementRef, ViewChild} from '@angular/core';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss']
 })
-export class AppComponent implements AfterViewInit {
-  @ViewChild('appTitle')
-  title!: ElementRef<HTMLHeadingElement>;
-
-  ngAfterViewInit(): void {
-      this.title.nativeElement.innerText = 'Shop';
-  }
-
+export class AppComponent {
 }