+++ src/app/feature/admin/admin-routing.config.ts
@@ -1,0 +1,7 @@
+export const adminRouting = {
+  products: 'products',
+  add: 'product/add',
+  edit: 'admin/product/edit',
+  order: 'order',
+}
+