+++ src/app/feature/cart-list/cart-list.module.ts
@@ -1,8 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CartListComponent } from './cart-list.component';
 import { SharedModule } from 'src/app/shared/shared.module';
-import { CartItemModule } from '../cart-item/cart-item.module';
 import { NgSelectModule } from '@ng-select/ng-select';
+import {CartListRoutingModule} from "./cart-list-routing.module";
 
 
 
@@ -10,9 +10,8 @@ import { NgSelectModule } from '@ng-select/ng-select';
   declarations: [CartListComponent],
   imports: [
     SharedModule,
-    CartItemModule,
     NgSelectModule,
+    CartListRoutingModule
   ],
-  exports: [CartListComponent]
 })
 export class CartListModule { }