+++ src/app/feature/process-order/process-order.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-process-order',
+  templateUrl: './process-order.component.html',
+})
+export class ProcessOrderComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit(): void {
+  }
+
+}