+++ src/app/shared/models/index.ts
@@ -1,0 +1,7 @@
+export * from './product.model'
+export * from './ng-select-types'
+export * from './config.model'
+export * from './cart-statistics'
+export * from './auth-config'
+export * from './app-info'
+export * from './route.config'