+++ src/app/shared/models/product.model.ts
@@ -1,11 +1,12 @@
-export interface ProductModel {
+  export interface ProductModel {
+    id: number,
     name: string,
     image: string,
     price: number,
-    isAvailable: boolean
+    isAvailable: boolean,
+    phoneInCart?: number
   }
 
   export interface CartProductModel extends ProductModel {
     phoneInCart: number
   }
-