+++ src/app/shared/models/route.config.ts
@@ -1,0 +1,7 @@
+export enum RouteConfig {
+  loginPage = 'login',
+  productsPage = 'products',
+  cartPage = 'cart',
+  adminPage = 'admin',
+  order = 'order'
+}